requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aiohttp
aiosignal
ansi
anyio
async-timeout
attrs
bcrypt
beautifulsoup4
black
certifi
cffi
charset-normalizer
click
colorlog
cryptography
daemonize
deepmerge
dulwich
errbot
errbot-backend-slackv3 @ git+https://github.com/errbotio/err-backend-slackv3.git@main
flask
frozenlist
h11
httpcore
httpx
idna
itsdangerous
jinja2
mariadb
markdown
MarkupSafe
multidict
mypy-extensions
packaging
paramiko
pathspec
platformdirs
pycparser
pyee
Pygments
pygments-markdown-lexer
pynacl
pyOpenSSL
python-dotenv
python-telegram-bot
requests
setuptools
six
slack-sdk
slackeventsapi
sniffio
soupsieve
sshtunnel
typing-extensions
urllib3
waitress
WebOb
webtest
Werkzeug
yarl
jira
python-jenkins
dotenv
